requirements.txt
python-dotenv
langchain
langchainhub
langchain-community
langchain-openai
langgraph
ollama
gradio #For UI

tavily-python
# need addtional support for the installation process
pygraphviz # https://pygraphviz.github.io/documentation/stable/install.html
aiosqlite # https://aiosqlite.omnilib.dev/en/latest/

notebook
pre-commit
